requirements.txt
pynput
pyautogui
pytesseract
openai
pyperclip
openrgb-python
Pillow
opencv-python
numpy